shell: refactor `tailor` goal into `goals` package
